icu: fix Windows Cygwin cross build
This replaces the previous, broken Windows ARM64 solution with a
general fix for Cygwin cross-building. The main problem is the
PATH environment, because that is colon-seperated on Cygwin, like
on Unix, so won't work with any absolute "Windows" path, which is
needed for the --with-cross-build option.

One general change is the adoption of icucross.inc. I don't know,
why that should prefer the general CURR_FULL_DIR from icudefs.mk
over the specific one in @platform_make_fragment@. That breaks
here, because it returns a cygwin unix path, which is used as
an option to compiled tools, which these won't be able to handle.

Change-Id: Ic2cf15b48801ac57ea5a8a2876ce59f2a84edfb3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103759
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>