sw/qa: relax testTdf134146 in layout tests.
The old test case expects the number of //text to
be 14. I.e. the Y-axis title splits into exactly two
lines. But it splits into 3 lines in my environment
( Windows10 traditional Chinese version ). Considering
the result might be locale and font dependent, relax the
test case so that it passes as long as the title breaks
into two or more lines.

Change-Id: Ic1da3993c95ecaa7a58d09901a7ce05dd8a34235
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103900
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>