Add a tool that will be useful for various tasks if/when we build LO on WSL
It is a Win32 program, not a Linux (WSL) one.

So far just one function: Convert a full Windows pathname to 8.3
format (if present). WSL does not seem to have that functionality
built-in, and doing it through some invocations of cmd.exe or a .bat
file from WSL seems very complex.

Change-Id: I4c29e2382b0230aacbd22ca6ea0814c8c4d01866
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103912
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>