don't disable the fast path in DrawDeviceAlphaBitmap() (tdf#137311)
Only OpenGL and Skia implement the BlendBitmap*() calls, and so I
disabled the direct path in an attempt to avoid forcing the mirroring
if it takes place. But there's also the DrawAlphaBitmap() call
that is implemented for other backends. So always do the mirroring
already there if it takes place.

Change-Id: I8c4c96ea18ac55ebad041e0d28c4228542d9b2e6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104206
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>