Qt5 std::abort, if we're unable to paint
If this crashes LO too often, we can revert it. It helped me to
debug tdf#137804. Then we might need to work with exceptions or
some other way to notifiy the creator of the Qt5Painter, that
the constructor has failed. Much more work to handle all sites
using Qt5Painter...

Change-Id: I81cd18770dc93d8dd1542972cd4fc220dd0dce09
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104894
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>