Resolves tdf#132479: Rename "equal triangle" to "regular triangle"
"Equal triangle" is not a correct English term, as people use "regular
triangle" or "equilateral triangle".

Technically this is not a true regular triangle (i.e. equilateral and
equiangular), but an isosceles triangle with equal width and height.
For the name of an animation path there is no need for too much detail
though, and "regular triangle" is close enough.

Change-Id: I72f5373ae68dc1ba771cf605073c15210c5300e7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105223
Tested-by: Jenkins
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>