tdf138138 Apply shape shift also to NoRotatedAnchor
The 'normal' anchor (as on screen) is connected to the snap rectangle,
the NoRotatedAnchor is connected to the logic rectangle. They differ,
if the shape is transformed, e.g. rotated. Error was, that values of
the 'normal' anchor were applied to NoRotatedAnchor instead of
calculating the shift of NoRotatedAnchor independently. The error
becomes only visible on save, because there the NoRotatedAnchor is
used. Effected shape types are legacy shapes, text boxes and
transformable OLEs.
I have not tested, whether this fix would work for LO 7.0 too.

Change-Id: I8ad22ca54bdd49861a16a34736860a9871d8eba0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105611
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>