Remove unused SfxFrameDescriptor::SetActualURL parameter
It got technically unused when e79e8117dcc7475d8d90afeaaac9eb7050ff244e
"loplugin:unusedfields in sfx2" removed SfxFrameDescriptor::aActualURL, which in
turn had apparently become unused over a series of commits cleaning up unused
code.  The function looks like a misnomer now, and it is not clear to me whether
what it does is still useful (and whether or not all the clean-up that brought
us here was actually good), but lets leave the remains as they are for now.

Change-Id: I0479068176cc72193b2b44fc49d8fd3d71eadcb5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106013
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>