tdf#138010 (II) getNativeControlRegion (gtk3/kf5): Consider frame's border
Take the frame border/padding into account when calculating
the content rectangle (i.e. the area that can safely be
drawn into without overlapping the actual frame) when
calculating dimensions in 'getNativeControlRegion' regardless
of the exact 'DrawFrameFlags' being passed.

Previously, gtk3 and qt5 only did this when the
'DrawFrameFlags::NoDraw' flag was passed in addition,
but the actual drawing routine 'drawNativeControl'
does not make any such distinction, so the frame ended
up at a place that was still in the "content rectangle"
as calculated previously (which was the same as the bounding
rect).

Returning the actual content rect is a prerequisite
for making 'VclScrolledWindow' take the width of a natively
drawn frame into account, which will be done in a
follow-up commit.

Change-Id: I8d3a3fbd387108a24a0478e3465c8950e6d59735
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106155
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>