remove non-functional resize grip at right of basic ide tabbar
left behind after....

commit 867af25bd8d36c7e5ef6f16c2876ac82e63648cc
Date:   Fri Jan 22 12:04:05 2016 +0100

    tdf#96683 Move tabs in Basic IDE to an own row

before that it could be used to assign sizes to the tabbar vs
scrollbar when they were on the same line. Now that they are on
different lines it shouldn't be there.

See tdf#87684 for the similar calc case where
Office::Calc::Layout::Other::TabbarInlineWithScrollbar exists to
re-enable this feature for calc.

Change-Id: I6430e13ab79f4c424ae35f7b8e3ad7bd47ec201b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106275
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>