tdf#138253 sw textbox: fix content of draw format on undo
Regression from commit da4f9b77a6cd39b1ae5babdd476d1575c8b9371c
(tdf#135149 sw: fix deleting textbox of as-char shapes, 2020-09-07), the
crash was caused by a user-after-free triggered from the X11 clipboard
code. This could happen beause the clipboard document had a draw frame
format, which has an SwNode pointer, but the SwNode instance was outside
that clipboard document, and the owning document is already gone. So by
the time the clipboard document would be deleted, the SwNodeIndex can't
de-register itself.

The root cause was that the doc model was corrupted after a cut of a
textbox + undo: the textbox pointers of the fly/draw formats were OK,
but not the SwFormatContent of the draw format.

Change-Id: I5761b72948caca397320aed801559e8493c33e1b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106453
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>