clean up enabling/disabling GUI options related to Skia
This should properly enable/disable 'force Skia software rendering'
and 'use hardware acceleration' checkboxes when Skia is enabled or
disabled.
Technically the two options are duplicates, since 'use hardware
acceleration' is the exact opposite of 'force Skia software
rendering'. But the implementation of the hw accel option is so tied
to the implementation of the canvas module that it's simpler
to ignore it for Skia.

Change-Id: Ia7c54140aec72d6ef75b7ba53ce83037311f0caf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106491
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>