graphic: clean-up switches GDIMetafile shouldn't be default
There are many switches ime ImpGraphic depending on the graphic
type. The default many times assumes to be GDIMetafile, which is
kind-of dangerous. Add all GraphicType cases to the switch (so
to handle "NONE" type if they weren't present) and add GdiMetaFile
type switch instead of default. No default is needed because all
the types are handled and if a new type is added, we will get an
warning - which is what we want.

Change-Id: I2ed5e7bce604c7215ce7101db9a04125c7a9244e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/107285
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>