tdf#61154 sw Table of Contents: remove tracked deletion
... part of a chapter, section etc. names also in Show
Changes mode, instead of showing as normal text.

Note: This problem was worse, then the case of references,
because hiding tracked changes didn't fix the Table of
Contents automatically, so it's better to avoid the
generation of the bad mixed names, which showed also
the deleted text as normal, without any visible differences
(strikethrough etc.).

Change-Id: Ieef8cf6e2f8bef64f99e5cb0542b7b43d8e76a27
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/107176
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>