tdf#134128 Use Gdiplus::DashCapRound for round dash or dot.
If Skia and OpenGL are disabled, rendering in edit mode shows no round
dashes and dots on Windows. This becomes especially visible, when
importing OOXML documents. In that case it looks as if dots are lost.
The patch uses now Gdiplus::DashCapRound instead of the previously
used Gdiplus::DashCapFlat, and it adds a similar tweak as in OOXML
import and increases the dash length by the cap size for rendering.

Change-Id: I98a258809ef253a2cacb7c5c94f2b26b89ee2488
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/107181
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>